cleanup(pkg/scrape): refactor files
