--- pkg/scrape/arch.go
@@ -1,7 +1,0 @@
-package scrape
-
-type Arch string
-
-const (
-	X86_64Arch Arch = "x86_64"
-)
