--- pkg/scrape/debug.go
@@ -1,7 +1,0 @@
-package scrape
-
-import "fmt"
-
-func debug(format string, a interface{}) {
-	fmt.Printf("\n" + format, a)
-}
