+++ pkg/scrape/defaults.go
@@ -1,6 +1,8 @@
 package scrape
 
 const (
+	X86_64Arch Arch = "x86_64"
+
 	// Default architecture for which scrape for packages
 	DefaultArch = X86_64Arch
