--- pkg/scrape/filter.go
@@ -1,4 +1,0 @@
-package scrape
-
-// A package filter string prefix
-type Filter string
