--- pkg/scrape/package.go
@@ -1,7 +1,0 @@
-package scrape
-
-type PackagesURIFormat string
-
-type PackagesURITemplate string
-
-type Package string
