--- pkg/scrape/repository.go
@@ -1,6 +1,0 @@
-package scrape
-
-type Repository struct {
-	Name                string
-	PackagesURITemplate PackagesURITemplate
-}
